fix(propagators): asymptotic-twin x64 guard + ij-meshgrid decompose + dtype/copy fixes + complexity-gate hardening (audit P3-51, P3-52, P3-53, P3-56, P3-57 + wave-5 follow-up)
- The four asymptotic JAX twins adopt the rcwa-style x64 guard and no
  longer mutate jax_enable_x64 MID-CALL; BEHAVIOR: they now RAISE when x64
  is off instead of flipping global config (P3-52).
- decompose_lg/decompose_hg derive dx/dy from the axis with variation, so
  indexing='ij' meshgrids work (previously silent all-zero coefficients)
  (P3-53).
- apply_fresnel_curvature honors dtype-follows-input (P3-51);
  fresnel_propagate drops an avoidable full-grid copy + a redundant astype
  (byte-identical) (P3-56); the dead chunk_output parameter is deprecated
  (P3-57).
- The pyFFTW gate itself (_fft2/_ifft2/_nd) now requires complex input so
  no future real-input caller can poison the bare-shape blacklist.

+++ lumenairy/propagators/fresnel.py
@@ -152,8 +152,16 @@ def fresnel_propagate(
     X2, Y2 = xp.meshgrid(x2, y2, indexing='xy')
 
     # -- quadratic phase in input plane --------------------------------------
-    phase_in = xp.exp(1j * k / (2 * z) * (X1**2 + Y1**2)).astype(target_cdtype)
-    E_mod = E_in.astype(target_cdtype) * phase_in
+    # v5.17.x (audit P3-56): astype(copy=False) throughout -- the phase
+    # screen is freshly allocated by xp.exp (never aliased) and E_in is
+    # only *read* by the multiply below, so neither needs the defensive
+    # copy.  Pre-fix the bare astype paid two avoidable full-grid copy
+    # passes (~4 GB transient each at 16384^2 complex128), partially
+    # undoing the v5.17.0 lifetime hygiene in this function.
+    # Byte-identical output.
+    phase_in = xp.exp(1j * k / (2 * z) * (X1**2 + Y1**2)).astype(
+        target_cdtype, copy=False)
+    E_mod = E_in.astype(target_cdtype, copy=False) * phase_in
     # v5.17.0 lifetime hygiene (byte-identical): the input-plane grids and
     # phase screen are consumed -- free before the FFT so they don't ride
     # through it.  (JAX arrays are immutable/functional; del is a no-op
@@ -171,7 +179,8 @@ def fresnel_propagate(
     prefactor = (xp.exp(1j * k * z) / (1j * wavelength * z)
                  * xp.exp(1j * k / (2 * z) * (X2**2 + Y2**2))
                  * dx * dy)
-    prefactor = prefactor.astype(target_cdtype)
+    # v5.17.x (audit P3-56): copy=False -- prefactor is freshly built above.
+    prefactor = prefactor.astype(target_cdtype, copy=False)
     del X2, Y2
 
     E_out = prefactor * E_fft
@@ -299,7 +308,8 @@ def fraunhofer_propagate(
     prefactor = (xp.exp(1j * k * z) / (1j * wavelength * z)
                  * xp.exp(1j * k / (2 * z) * (X2**2 + Y2**2))
                  * dx * dy)
-    prefactor = prefactor.astype(target_cdtype)
+    # v5.17.x (audit P3-56): copy=False -- prefactor is freshly built above.
+    prefactor = prefactor.astype(target_cdtype, copy=False)
 
     E_out = prefactor * E_fft
